+++ lib/decorators/index.ts
@@ -1,5 +1,3 @@
-export * from './before' 
-export * from './after' 
 export * from './validate' 
 export * from './has-permission' 
 export * from './routes'