+++ lib/index.ts
@@ -1,3 +1,3 @@
 export * from './declarations'
-export * from './hooks'
+export * from './decorators'
 export * from './service'