+++ tests/implicit-routes.spec.ts
@@ -25,10 +25,10 @@ describe('Implicit routes', () => {
   it('should handle an implicit index GET route', (done: Function) => {
     chai
       .request(server)
-      .get('/users/1')
+      .get('/users')
       .end((err, res) => {
         expect(res).to.have.status(200)
-        expect(res.body).to.have.property('user')
+        expect(res.body).to.have.property('users').with.lengthOf(1)
         done()
       })
   })
@@ -36,10 +36,10 @@ describe('Implicit routes', () => {
   it('should handle an implicit single GET route', (done: Function) => {
     chai
       .request(server)
-      .get('/users')
+      .get('/users/1')
       .end((err, res) => {
         expect(res).to.have.status(200)
-        expect(res.body).to.have.property('users').with.lengthOf(2)
+        expect(res.body).to.have.property('user')
         done()
       })
   })